[hotfix] Extract JobLeaderService interface + rename impl into DefaultJobLeaderService
This commit allows to provide different JobLeaderService implementations to the TaskExecutor.
